Create QualtricsVideoRecorder
This is an example of a video recorder that can be implemented directly into the HTML of Qualtrics. This video recorder allows participants to take a recording of themselves and then send their audio file through file.io (you will need to make an account if you want to use this method). Note that this method does not collect video data in the interest of participant anonymity.